python/requirements.txt
locust
granian
multiprocessing
orjson
aiomysql
python-dotenv
uvicorn
asyncmy
fastapi